Add DrawingMarks component Add DrawingMarks to render a collection of marks, and handle deleting individual marks. Observe DrawingMarks from the InteractionLayer to re-render when we move, resize or delete marks.
