+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/InteractionLayer/components/DrawingMarks/index.js
@@ -1,0 +1,1 @@
+export { default } from './DrawingMarks'
